Move configloader imports back to root
