+++ core/src/modules/configLoaders/property.ts
@@ -7,8 +7,7 @@ import {
   validateConfigObjectKeys,
   IdsByClass,
 } from "../../classes/codeConfig";
-import { Source } from "../../models/Source";
-import { Property } from "../../models/Property";
+import { Source, Property } from "../.."; // configLoader imports need to be from root
 import { FilterHelper } from "../filterHelper";
 import { Op } from "sequelize";
 import { Deprecation } from "../deprecation";
