+++ core/src/modules/configLoaders/teamMember.ts
@@ -6,8 +6,7 @@ import {
   logModel,
   validateConfigObjectKeys,
 } from "../../classes/codeConfig";
-import { Team } from "../../models/Team";
-import { TeamMember } from "../../models/TeamMember";
+import { Team, TeamMember } from "../.."; // configLoader imports need to be from root
 import { Op } from "sequelize";
 
 export async function loadTeamMember(
